put the lib files in a .mlogjs folder
This makes it more clear to the user that these files are not meant
to be tracked by git, as well as making it easier to perform updates to
these library declaration files in the future
